requirements/requirements.txt
py-cord
ezcord
easy_pil
#mariadb
icecream
pyyaml
tornado
jinja2
chat_exporter
openai
python-dotenv
selenium
thefuzz
cryptography